update - visuals health/armor fix - visuals sound not using events - visual plots rewritten without ex rendering window struct - counters completely removed, using accurate round stats shared by server to client and list of impacts as all shots counter - prediction fixes, not fully done. Still have to search method to find what's wrong with weapon inaccuracy. - visual radar fixes, better I guess? - small change for hooked functions definitions - the return type is gathered directly from the struct. - couple of server stuff, wip to remove all events and use several hooks that server already listens and does what we need, great example is in hitmarker as comment. - small change for handles, make this as a struct wrapping the CBaseHandle field. - wip chams editor

+++ cheats/features/visuals/player/player.hpp
@@ -38,8 +38,8 @@ private:
 	void resetDormacy([[maybe_unused]] IGameEvent* event);
 	void updateDormacy(Player_t* ent);
 
-	std::array<int, 65> m_health;
-	std::array<int, 65> m_armor;
+	std::array<float, 65> m_health;
+	std::array<float, 65> m_armor;
 
 	struct DormacyInfo_t
 	{
